requirements.txt
apache-airflow-providers-apprise==1.2.2
apache-airflow-providers-amazon==7.0.0